# server/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos巨石应用实践API
    description: Kratos巨石应用实践API
    contact:
        name: bugsmo
        url: https://github.com/bugsmo/chunyu
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/bugsmo/chunyu/blob/master/LICENSE
    version: "1.0"
paths:
    /admin/v1/dict:details:
        get:
            tags:
                - DictDetailService
            description: 查询字典详情列表
            operationId: DictDetailService_ListDictDetail
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
                - name: fieldMask
                  in: query
                  description: 字段掩码
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListDictDetailResponse'
        post:
            tags:
                - DictDetailService
            description: 创建字典详情
            operationId: DictDetailService_CreateDictDetail
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateDictDetailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/dict:details/{detail.id}:
        put:
            tags:
                - DictDetailService
            description: 更新字典详情
            operationId: DictDetailService_UpdateDictDetail
            parameters:
                - name: detail.id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DictDetail'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/dict:details/{id}:
        get:
            tags:
                - DictDetailService
            description: 查询字典详情
            operationId: DictDetailService_GetDictDetail
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: code
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DictDetail'
        delete:
            tags:
                - DictDetailService
            description: 删除字典详情
            operationId: DictDetailService_DeleteDictDetail
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/dicts:
        get:
            tags:
                - DictService
            description: 查询字典列表
            operationId: DictService_ListDict
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
                - name: fieldMask
                  in: query
                  description: 字段掩码
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListDictResponse'
        post:
            tags:
                - DictService
            description: 创建字典
            operationId: DictService_CreateDict
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateDictRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/dicts/{dict.id}:
        put:
            tags:
                - DictService
            description: 更新字典
            operationId: DictService_UpdateDict
            parameters:
                - name: dict.id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Dict'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/dicts/{id}:
        get:
            tags:
                - DictService
            description: 查询字典
            operationId: DictService_GetDict
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Dict'
        delete:
            tags:
                - DictService
            description: 删除字典
            operationId: DictService_DeleteDict
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
        CreateDictDetailRequest:
            type: object
            properties:
                detail:
                    $ref: '#/components/schemas/DictDetail'
                operatorId:
                    type: integer
                    format: uint32
            description: 创建字典详情 - 请求
        CreateDictRequest:
            type: object
            properties:
                dict:
                    $ref: '#/components/schemas/Dict'
                operatorId:
                    type: integer
                    format: uint32
            description: 创建字典 - 请求
...
        Dict:
            type: object
            properties:
                id:
                    type: integer
                    description: ID
                    format: uint32
                name:
                    type: string
                    description: 字典名称
                description:
                    type: string
                    description: 描述
                creatorId:
                    type: integer
                    description: 创建者ID
                    format: uint32
                creatorName:
                    type: integer
                    description: 创建者名字
                    format: uint32
                createTime:
                    type: string
                    format: date-time
                updateTime:
                    type: string
                    format: date-time
                deleteTime:
                    type: string
                    format: date-time
            description: 数据字典
        DictDetail:
            type: object
            properties:
                id:
                    type: integer
                    description: 字典详情ID
                    format: uint32
                dictId:
                    type: integer
                    description: 字典ID
                    format: uint32
                orderNo:
                    type: integer
                    description: 排序号
                    format: int32
                label:
                    type: string
                    description: 字典标签
                value:
                    type: string
                    description: 字典值
                creatorId:
                    type: integer
                    description: 创建者ID
                    format: uint32
                creatorName:
                    type: integer
                    description: 创建者名字
                    format: uint32
                createTime:
                    type: string
                    format: date-time
                updateTime:
                    type: string
                    format: date-time
                deleteTime:
                    type: string
                    format: date-time
            description: 数据字典详情
...
        ListDictDetailResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/DictDetail'
                total:
                    type: integer
                    format: int32
            description: 查询字典详情列表 - 答复
        ListDictResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Dict'
                total:
                    type: integer
                    format: int32
            description: 查询字典列表 - 答复
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...